+++ lib/generators/boring/twilio/install/templates/twilio.rb
@@ -1,0 +1,4 @@
+Twilio.configure do |config|
+  config.account_sid = "TWILIO_ACCOUNT_SID"
+  config.auth_token = "TWILIO_AUTH_TOKEN"
+end